% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_normals.R
\name{climate_normals}
\alias{climate_normals}
\title{Climate normals}
\usage{
climate_normals(station)
}
\arguments{
\item{station}{String. The weather station ID.}
}
\value{
Average monthly weather.
}
\description{
Climate normals
}
\examples{
climate_normals(station = "10637")
}
