% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_data.R
\name{daily_data}
\alias{daily_data}
\title{Daily data}
\usage{
daily_data(station, start, end)
}
\arguments{
\item{station}{String. The weather station ID.}

\item{start}{String. The start date of the query (format: YYYY-MM-DD).}

\item{end}{String. The end date of the query (format: YYYY-MM-DD).}
}
\value{
Historical daily weather.
}
\description{
Daily data
}
\examples{
daily_data(station = "68816", start = "2019-06-01", end = "2019-06-30")
}
