% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_data.R
\name{hourly_data}
\alias{hourly_data}
\title{Hourly data}
\usage{
hourly_data(station, start, end, tz = "UTC", model = 0)
}
\arguments{
\item{station}{String. The weather station ID.}

\item{start}{String. The start date of the query (format: YYYY-MM-DD).}

\item{end}{String. The end date of the query (format: YYYY-MM-DD).}

\item{tz}{String. The time zone according to the tz database.}

\item{model}{Integer. A boolean parameter that specifies whether missing observations should be substituted with statistically optimized model data.}
}
\value{
Historical hourly weather.
}
\description{
Hourly data
}
\examples{
hourly_data(station = "10637", start = "2020-02-01", end = "2020-02-04")
}
