% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_station.R
\name{metadata_station}
\alias{metadata_station}
\title{Meta data}
\usage{
metadata_station(id = NULL, wmo = NULL, icao = NULL)
}
\arguments{
\item{id}{String. The Meteostat ID of the weather station.}

\item{wmo}{Integer. The WMO ID of the weather station.}

\item{icao}{String. The ICAO ID of the weather station.}
}
\value{
Meta data weather station.
}
\description{
Meta data
}
\examples{
metadata_station(id = "71508")
}
