% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby_stations.R
\name{nearby_stations}
\alias{nearby_stations}
\title{Nearby weather stations}
\usage{
nearby_stations(lat, lon, limit = 8, radius = 100)
}
\arguments{
\item{lat}{Numeric. The latitude of the location.}

\item{lon}{Numeric. The longitude of the location.}

\item{limit}{Integer. The maximum number of results.}

\item{radius}{Numeric. The radius of the query in kilometers.}
}
\value{
Available weather stations in the nearby of **lat** and **lon**.
}
\description{
Nearby weather stations
}
\examples{
nearby_stations(lat = 1.28, lon = 103.83, limit = 3)
}
