% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_climate_normals.R
\name{point_climate_normals}
\alias{point_climate_normals}
\title{Point climate normals}
\usage{
point_climate_normals(lat, lon, alt = 0)
}
\arguments{
\item{lat}{Numeric. The latitude of the geographic location.}

\item{lon}{Numeric. The longitude of the geographic location.}

\item{alt}{Integer. The elevation of the geographic location.}
}
\value{
Average monthly weather.
}
\description{
Point climate normals
}
\examples{
point_climate_normals(lat = -33.87, lon = 151.21, alt = 58)
}
