% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_daily_data.R
\name{point_daily_data}
\alias{point_daily_data}
\title{Point daily data}
\usage{
point_daily_data(lat, lon, alt = 0, start, end)
}
\arguments{
\item{lat}{Numeric. The latitude of the geographic location.}

\item{lon}{Numeric. The longitude of the geographic location.}

\item{alt}{Integer. The elevation of the geographic location.}

\item{start}{String. The start date of the query (format: YYYY-MM-DD).}

\item{end}{String. The end date of the query (format: YYYY-MM-DD).}
}
\value{
Historical daily weather.
}
\description{
Point daily data
}
\examples{
point_daily_data(lat = 33.749, lon = -84.388, alt = 336, start = "2019-06-01", end = "2019-06-30")
}
