% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_hourly_data.R
\name{point_hourly_data}
\alias{point_hourly_data}
\title{Point hourly data}
\usage{
point_hourly_data(lat, lon, alt = 0, start, end, tz = "UTC")
}
\arguments{
\item{lat}{Numeric. The latitude of the geographic location.}

\item{lon}{Numeric. The longitude of the geographic location.}

\item{alt}{Integer. The elevation of the geographic location.}

\item{start}{String. The start date of the query (format: YYYY-MM-DD).}

\item{end}{String. The end date of the query (format: YYYY-MM-DD).}

\item{tz}{String. The time zone according to the tz database.}
}
\value{
Historical hourly weather.
}
\description{
Point hourly data
}
\examples{
point_hourly_data(lat = 49.4967, lon = 8.4795, alt = 104, start = "2020-01-01", end = "2020-01-01", tz = "Europe/Berlin")
}
